overlay: migrate existing systems to OCI
In [1] we moved newly provisioned systems to be
deployed via container and thus retrieving updates
via zincati from the OCI registry (quay.io/fedora/fedora-coreos).

This starts the migration script shipped in [2]
before Zincati to set up the migration at the next
update.

Slightly tweak the migration script to look for
a stamp to opt-out of the migration and create one
to signal failure.

Also add a MOTD that look for those stamps and
display an appropriate message.

[1] https://github.com/coreos/fedora-coreos-tracker/issues/1823
[2] https://github.com/coreos/fedora-coreos-config/pull/3355

Closes https://github.com/coreos/fedora-coreos-tracker/issues/1890