some riscv enablement
There is an effort to build all of Fedora using the RISCV architecture.
This comment adds the yum repo files and also conditionalizes some
manifest items to make it possible to build for `riscv64`.

There are still some other pieces that need to land before builds
will succeed (like bootupd support for riscv64), but this is a start
in the right direction.

xref: https://github.com/coreos/fedora-coreos-tracker/issues/1931