coreos-boot-mount-generator: honor `boot=` karg better in multipath case
As described in https://github.com/openshift/os/issues/1787, we need to
loosen up our restriction that only a single `boot` and `root` labeled
filesystem exists, at least past the first boot to start.

There is code already that injects `boot` and `root` kargs to specify
the exact UUID we want so this is feasible.

There's currently two ways in which the boot device can be specified:
- the `/run/coreos/bootfs_uuid` file exists; this is written by rdcore
  but is only available on first boot
- a `boot=` karg, which supports the "dracut convention" (i.e.
  `boot=LABEL=` or `boot=UUID=`, etc.), exists; this is usually written
  by rdcore but is only available on subsequent boots

If neither of those are specified, then we just default to
`by-label/boot`.

But currently, if multipath is enabled, we override all that and just
always use `by-label/dm-mpath-boot`. This is not great, because a UUID
might've been specified, and that's much stronger than a label. It also
directly conflicts with the mentioned higher level goal of supporting
multiple `boot` labels.

Instead, rework multipath support so that it respects both `boot=` kargs
and `bootfs_uuid` files. When possible, we use the multipathed version
of the specified UUID/label written by our udev rules because it's less
racy. Otherwise, we just use the non-multipathed one.