tests/platforms: add test for Azure NVMe udev rules
Introduce a new test which verifies that udev rules[1] for Azure
Managed NVMe disks correctly create symlinks in `/dev/disk/azure`.
It only runs on Azure and uses new kola functionality[2] to override
the instance type, enabling the use of a Hyper-V Gen2 VM
(Standard_M16bds_v3) with a UltraSSD_LRS (NVMe) data disk attached.

The test checks for the presence of OS and data disk symlinks, and
that they point to a valid target.

[1]: https://github.com/coreos/fedora-coreos-config/pull/3378
[2]: https://github.com/coreos/coreos-assembler/pull/4109