Containerfile: use build arg files
buildah supports `--build-arg-file` which takes a file that defines
build arguments. This gives us a cleaner way to implement the semantics
we want here.

Pivot to that and drop all the `Containerfile.*` bits.

Semi-controversial: the default build arg values in the Containerfile
are purposely invalid so that everyone is forced to use the same args
when building the canonical image. This is more cumbersome but ensures
consistency. E.g. no default would make sense both on `testing-devel`
and `rawhide`.