build-rootfs: temporarily inject our passwd/group files
We want to keep our passwd/group files as canonical for now to eliminate
any changes in our currently fixed UIDs/GIDs as part of the rebase to
the bootc base images.

Obviously not ideal. The medium-term fix for this is to convert our
passwd and group files to sysusers entries instead, and then force
bootc-base-imagectl to treat sysusers as canonical.

See also: https://github.com/coreos/rpm-ostree/pull/5427