build-rootfs: add workaround for RHEL-97826
bootc owns files in `/usr/share/doc` which are correctly marked as
documentation, but because of a bug in rpm-ostree, the subdirectories
are still installed even with `documentation: false`. This throws off
`bootc container lint`.

This is fixed in newer bootc, but not RHEL 9.6.

Work around this there by checking if documentation is turned off
whether the issue is fixed, and if not, just "pass through" the docs
into the installed target via another overlay. This then will only
activate when building RHCOS 9.6. (Also c9s currently until bootc is
rebased there.)