konflux: onboard all the streams
We are defining the pipelines for all the streams in testing-devel
branch. The files will be propagated to 'next-devel', 'branched' and
'rawhide' by the coreos bot through push method. For the 'testing',
'next' and 'stable' branches, the files will be propagated during
the release procedure through PR method.

Defining all of them in 'testing-devel' avoid us to apply similar
changes by opening a PR against each respective source branch.
The downside is that only 'testing-devel' pipeline changes will be
truly tested, because the PR will be submitted against its branch.
But as we don't expect any divergence between all the streams in
Konflux perspective, I'd say it's ok. And if we want to test something
specific for a stream (i.e not testing-devel) before merging against
testing-devel, we still can open a draft PR against its respective
branch to trigger a Konflux pipeline and adjust the main PR if needed.

Also, we are using kustomize to generate the pipelinerun configuration
files based on the base one and with kustomize transformers. This will
avoid us having to replicate same change at different location which
is prone to errors.