tests/upgrade.extended: cleanup rollback before final deploy
Since we have to download and then deploy the final ociarchive we are
using more space than we normally would on a system so we are starting
to run out of free space when importing the ociarchive.

Let's just workaround this by dropping the rollback deployment since
that's an easy way to free up some space here.

Fixes https://github.com/coreos/fedora-coreos-tracker/issues/1997