tree: move repo files to use GPG keys in /etc/pki/rpm-gpg
The use of distribution-gpg-keys for these repo files is
ancient; the first occurrence is 342e8813 ("Switch to f29, add
fedora-release-coreos") which doesn't have a rationale for it.

However, in practice today FCOS is currently built only by Fedora-based
tools, like cosa and the bootc-base-imagectl in fedora-bootc. So we
could just use the `/etc/pki/rpm-gpg` paths there.

This allows us to avoid having to install `distribution-gpg-keys` before
doing a compose. Note that `fedora-rawhide.repo` was already using the
`/etc` path so we know this works.