tests: use TERM=dumb to avoid noise in machinectl shell output
As reported in [1], 'machinectl shell' in systmed-258-rc2 now writes
additional OSC stuff to stdout, breaking the tests. This might or
might not be an error [2], but while this is being figured out, use
TERM=dumb as a work-around. Also, add '-q' everywhere to trim output
about the transient unit name which is not interesting.

Fixes https://github.com/coreos/fedora-coreos-tracker/issues/2003.

[1] https://github.com/coreos/fedora-coreos-tracker/issues/2003
[2] https://github.com/systemd/systemd/issues/38527