coreos-bootc-minimal-plus: stop patching prepare-root.conf
I think the issue that motivated this hack was fixed by
https://github.com/ostreedev/ostree/pull/3316.

However, a new issue happens in our live boots when we do this,
which is that because ostree-prepare-root still remounts `/etc` with
`sysroot.readonly`, systemd sees this and auto-generates a mount unit
from that and that shadows _our_ generated `sysroot-etc.mount` which
mounts from the transient XFS loopback hack.

We should be able to tweak ostree-prepare-root for this, but for now
we still need to neuter `prepare-root.conf` in live boots. I think it's
cleaner though to do this directly in `ostree-cmdline.sh` rather than in
a postprocessing script to limit the hacks to just where it's needed and
also so that it transparently works in the container-native case, where
our postprocessing scripts run _after_ dracut. (We could rerun dracut,
but I'm trying to avoid that.)

Another consequence of doing it this way is that this should also allow
us to stop setting `sysroot.readonly` at osbuild time for the disk
images[[1]], and _unsetting_ it at osbuild time for the live media[[2]].

[1]: https://github.com/coreos/coreos-assembler/blob/3615165d645bb7a6f4cd8d004df44e427e2cc1d7/src/osbuild-manifests/coreos.osbuild.x86_64.mpp.yaml#L177
[2]: https://github.com/osbuild/osbuild/blob/67f344fa52c0942a9a4878317458a92cd20dc273/stages/org.osbuild.coreos.live-artifacts.mono#L587