Add concept of a `buildroot-prep` script
We want to fast-track the backport of
https://github.com/coreos/rpm-ostree/pull/5475 to unblock switch f44
and f43 over to the buildah path. Since this is likely not the first
time we'll do this and likely not only for FCOS, add a new concept of
an optional `buildroot-prep` script which can do $things like updating
packages before entering build-rootfs.

That specific script added here will not be automatically inherited by
coreos/rhel-coreos-config.