versionary: remove mechanical streams from unlocked streams
In the process of using konflux to build rawhide, we enabled lockfiles
for mechanial streams in [1].
Reflect that change in versionary since rawhide and branched need to be
locked streams. bodhi-updates* should technically also be locked but these
haven't been built in years so i won't touch them for now as they require
are more thorough cleanup. [3]
See [2] for more details.

[1] https://github.com/coreos/fedora-coreos-pipeline/pull/1256
[2] https://github.com/coreos/fedora-coreos-tracker/issues/2038#issuecomment-3498258143
[3] https://github.com/coreos/fedora-coreos-tracker/issues/1734