manifests: drop all keys that no longer have meaning
In a derived container build world, which we've switched to in F43+ [1],
there are a lot of historical RPM-OSTree manifest keys that are essentially
dead configuration (has no effect). If anything isn't explicitly handled
in `build-rootfs` like `recommends:` and `ostree-layers:` then it's
either not important or being handled a different way (like in a
postprocess).

[1] https://github.com/coreos/fedora-coreos-tracker/issues/1969